Chore- Updated Repository for dishes

+++ carbonO/carbonTracker/src/main/java/com/carbonO/Ingredient/Ingredient.java
@@ -27,9 +27,9 @@ public class Ingredient {
     private String ingredientName;
     @OneToMany(mappedBy = "ingredient", cascade = CascadeType.ALL)
     private List<DishRecipe> dishRecipe;
-    private Integer carbonFootprint;
+    private Double carbonFootprint;
 
-    public Ingredient(String ingredientName, Integer carbonFootprint) {
+    public Ingredient(String ingredientName, Double carbonFootprint) {
         this.ingredientName = ingredientName;
         this.carbonFootprint = carbonFootprint;
     }
